+++ src/k8s/cmd/k8sd/k8sd.go
@@ -33,9 +33,8 @@ func addCommands(root *cobra.Command, group *cobra.Group, commands ...*cobra.Com
 
 func NewRootCmd(env cmdutil.ExecutionEnvironment) *cobra.Command {
 	cmd := &cobra.Command{
-		Use:    "k8sd",
-		Short:  "Canonical Kubernetes orchestrator and clustering daemon",
-		PreRun: cmdutil.HookCheckLXD(true),
+		Use:   "k8sd",
+		Short: "Canonical Kubernetes orchestrator and clustering daemon",
 		Run: func(cmd *cobra.Command, args []string) {
 			// configure logging
 			log.Configure(log.Options{
